Move elements to ltl_tools folder
