+++ ltl_automaton_core/src/ltl_tools/buchi.py
