+++ ltl_automaton_core/src/ltl_tools/discrete_plan.py
