+++ ltl_automaton_core/src/ltl_tools/ltl2ba.py
