+++ ltl_automaton_core/src/ltl_tools/ltl_planner.py
