+++ ltl_automaton_core/src/ltl_tools/product.py
