+++ ltl_automaton_core/src/ltl_tools/promela.py
