+++ ltl_automaton_core/src/ltl_tools/ts.py
